server/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
python-multipart==0.0.18
openai==1.58.1
faster-whisper==1.1.0
onnxruntime==1.20.1
numpy==1.26.4
soundfile==0.12.1
websockets==14.1
